Began refactoring and modularizing js
