+++ public/index.html
@@ -3,22 +3,12 @@
     <title>W2 Helper</title>
     <script src='./js/vendor/jquery-3.2.1.js'></script>
     <script src='./js/vendor/jsrender.js'></script>
-    <script src='./js/templates/employee.js'></script>
-    <script src='./js/templates/business.js'></script>
-    <script src='./js/w2helper.js'></script>
+    <script src='./js/app.js'></script>
     <link rel='stylesheet' type='text/css' href='./css/vendor/bootstrap.min.css'></style>
     <link rel='stylesheet' type='text/css' href='./css/w2helper.css'></style>
   </head>
   <body>
     <main>
-      <div class='business'></div>
-      <div id='employees'></div>
-      <div class='controls'>
-        <button id='add-employee' class='btn btn-default'>Add Employee</button>
-        <button id='print-report' class='btn btn-default'>Print Report</button>
-      </div>
-      <div id='printout'>
-      </div>
     </main>
   </body>
 </html>
