+++ public/js/w2helper.js
@@ -1,3 +1,5 @@
+var Report = require('./models/report');
+
 $(document).ready(function() {
   initialize();
 });
@@ -6,75 +8,11 @@ HEADERS = 'BusinessName, BusinessFEIN, BusinessAddress, BusinessCity, BusinessSt
 BUSINESSFIELDS = ['legal-name', 'business-FEIN', 'business-address', 'business-city', 'business-state', 'business-zip', 'business-withholdingAcct'];
 EMPLOYEEFIELDS = ['employee-firstname', 'employee-lastname', 'employee-SSN', 'employee-address', 'employee-city', 'employee-state', 'employee-zip', 'employee-wages'];
 
-var employeeCount = 0;
+var nextReportID = 0;
 
 function initialize() {
-  $('#print-report').on('click', printReport);
-  $('#add-employee').on('click', addEmployee);
-  $('#employees').on('click', '.remove-employee', removeEmployee);
-  $('.business').html($.render.business());
-}
-
-function printReport() {
-  var lines = [];
-  var businessObject = serializeFormToObject('business');
-  var businessLineData = '';
-  var employeeObjects = $('form.employee').map(function(i,d) {
-    return serializeFormToObject(d.id);
-  });
-
-  lines.push(HEADERS);
-
-  BUSINESSFIELDS.forEach(function(fieldName) {
-    businessLineData += businessObject[fieldName] + ', ';
-  });
-
-  employeeObjects.each(function(index, employee) {
-    var employeeLineData = '';
-    console.log(employee);
-    EMPLOYEEFIELDS.forEach(function(fieldName) {
-      employeeLineData += employee[fieldName] + ', ';
-    });
-    lines.push(businessLineData + employeeLineData.substring(0, employeeLineData.lastIndexOf(',')));
-  });
-
-  displayData(lines);
-}
-
-function serializeFormToObject(id) {
-  var data = $('#'+id).serializeArray();
-  var mapped = {};
-  data.forEach(function(d) {
-    mapped[d.name] = d.value;
-  });
-  return mapped;
-}
-
-function generateLine(business, employee) {
-  BUSINESSFIELDS.forEach()
-}
-
-function displayData(lines) {
-  var html = "<table>"
-  lines.forEach(function(row) {
-    html += "<tr>";
-    row.split(', ').forEach(function(cell) {
-      html += "<td>";
-      html += cell;
-      html += "</td>";
-    });
-    html += "</tr>";
-  });
-  html += "</table>";
-
-  $('#printout').html(html);
-}
-
-function addEmployee() {
-  $('#employees').append($.render.employee({ id: employeeCount }));
-  employeeCount++;
-}
-
-function removeEmployee(e) {
-  $(e.target).closest('.form').remove();
+  var reportContext = $("main");
+  var currentReport = Report(reportContext, nextReportID);
+  currentReport.render();
+  nextReportID += 1;
 }
